refactor(access): move function get_parent_obj to common viewset
ref: #982 #980
